+++ src/frontend/src/sol/components/fee/SolFeeContext.svelte
@@ -1,10 +1,14 @@
 <script lang="ts">
-	import { assertNonNullish } from '@dfinity/utils';
+	import { assertNonNullish, nonNullish } from '@dfinity/utils';
 	import { getContext, onDestroy } from 'svelte';
 	import { i18n } from '$lib/stores/i18n.store';
 	import { SEND_CONTEXT_KEY, type SendContext } from '$lib/stores/send.store';
 	import { replacePlaceholders } from '$lib/utils/i18n.utils';
-	import { estimatePriorityFee } from '$sol/api/solana.api';
+	import {
+		estimatePriorityFee,
+		getSolCreateAccountFee,
+		loadTokenAccount
+	} from '$sol/api/solana.api';
 	import {
 		MICROLAMPORTS_PER_LAMPORT,
 		SOLANA_TRANSACTION_FEE_IN_LAMPORTS
@@ -12,10 +16,12 @@
 	import { SOL_FEE_CONTEXT_KEY, type FeeContext } from '$sol/stores/sol-fee.store';
 	import { mapNetworkIdToNetwork } from '$sol/utils/network.utils';
 	import { isTokenSpl } from '$sol/utils/spl.utils';
+	import { isNullishOrEmpty } from '$lib/utils/input.utils';
 
 	export let observe: boolean;
+	export let destination = '';
 
-	const { feeStore, prioritizationFeeStore }: FeeContext =
+	const { feeStore, prioritizationFeeStore, ataFeeStore }: FeeContext =
 		getContext<FeeContext>(SOL_FEE_CONTEXT_KEY);
 
 	const { sendToken, sendTokenNetworkId } = getContext<SendContext>(SEND_CONTEXT_KEY);
@@ -57,6 +63,39 @@
 	const clearTimer = () => clearInterval(timer);
 
 	onDestroy(clearTimer);
+
+	const updateAtaFee = async () => {
+		if (isNullishOrEmpty(destination) || !isTokenSpl($sendToken)) {
+			ataFeeStore.setFee(undefined);
+			return;
+		}
+
+		const solNetwork = mapNetworkIdToNetwork($sendTokenNetworkId);
+
+		assertNonNullish(
+			solNetwork,
+			replacePlaceholders($i18n.init.error.no_solana_network, {
+				$network: $sendTokenNetworkId.description ?? ''
+			})
+		);
+
+		const tokenAccount = await loadTokenAccount({
+			address: destination,
+			network: solNetwork,
+			tokenAddress: $sendToken.address
+		});
+
+		if (nonNullish(tokenAccount)) {
+			ataFeeStore.setFee(undefined);
+			return;
+		}
+
+		const ataFee = await getSolCreateAccountFee(solNetwork);
+
+		ataFeeStore.setFee(ataFee);
+	};
+
+	$: destination, $sendToken, updateAtaFee();
 </script>
 
 <slot />
